He creado la clase abstracta Indice, la cual será la plantilla para mi nueva clase IndiceOrdenado
